% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score_Player.R
\name{Score_Player}
\alias{Score_Player}
\title{A NBA Player Function}
\usage{
Score_Player(x)
}
\description{
This function allows you to select the player who scored the most points in a specific year.
}
\examples{
Score_Player(1991)
}
\keyword{most}
\keyword{points}
